+++ controller/message.go
@@ -15,7 +15,7 @@ type MessageController struct{}
 type MessageActionRequest struct {
 	Token      string `query:"token"              vd:"len($)>0"` // 用户鉴权token
 	ToUserID   int64  `query:"to_user_id,string"  vd:"$>0"`      // 对方用户id
-	ActionType int    `query:"action_type,string" vd:"$==1"`     // 1-发送消息
+	ActionType int64    `query:"action_type,string" vd:"$==1"`     // 1-发送消息
 	Content    string `query:"content"            vd:"len($)>0"` // 消息内容
 }
