+++ service/message.go
@@ -7,7 +7,7 @@ import (
 	"github.com/cloudwego/hertz/pkg/common/hlog"
 )
 
-func MessageAction(userID, toUserID int64, actionType int, content string) (*response.MessageActionResponse, error) {
+func MessageAction(userID, toUserID int64, actionType int64, content string) (*response.MessageActionResponse, error) {
 	// 操作数据库
 	err := mysql.MessageAction(userID, toUserID, content)
 	if err != nil {
